onDelete moved into GetLifeCycle
