+++ lib/get_state_manager/src/rx_flutter/rx_disposable.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 import '../../../get_instance/src/lifecycle.dart';
 
@@ -15,14 +16,8 @@ abstract class DisposableInterface with GetLifeCycle {
   /// Checks whether the controller has already been initialized.
   bool get initialized => _initialized;
 
-  bool _isClosed = false;
-
-  /// Checks whether the controller has already been closed.
-  bool get isClosed => _isClosed;
-
   DisposableInterface() {
     onStart.callback = _onStart;
-    onDelete.callback = _onDelete;
   }
 
   // Internal callback that starts the cycle of this controller.
@@ -33,22 +28,17 @@ abstract class DisposableInterface with GetLifeCycle {
     SchedulerBinding.instance?.addPostFrameCallback((_) => onReady());
   }
 
-  // Internal callback that starts the cycle of this controller.
-  void _onDelete() {
-    if (_isClosed) return;
-    _isClosed = true;
-    onClose();
-  }
-
   /// Called immediately after the widget is allocated in memory.
   /// You might use this to initialize something for the controller.
   @override
-  void onInit() {}
+  @mustCallSuper
+  void onInit() => super.onInit();
 
   /// Called 1 frame after onInit(). It is the perfect place to enter
   /// navigation events, like snackbar, dialogs, or a new route, or
   /// async request.
   @override
+  @mustCallSuper
   void onReady() {}
 
   /// Called before [onDelete] method. [onClose] might be used to
@@ -58,5 +48,6 @@ abstract class DisposableInterface with GetLifeCycle {
   /// like TextEditingControllers, AnimationControllers.
   /// Might be useful as well to persist some data on disk.
   @override
+  @mustCallSuper
   void onClose() {}
 }
