+++ lib/get_state_manager/src/rx_flutter/rx_notifier.dart
@@ -39,7 +39,6 @@ typedef Condition = bool Function();
 abstract class GetNotifier<T> extends Value<T> with GetLifeCycle {
   GetNotifier(T initial) : super(initial) {
     onStart.callback = _onStart;
-    onDelete.callback = _onDelete;
     _fillEmptyStatus();
   }
 
@@ -48,24 +47,12 @@ abstract class GetNotifier<T> extends Value<T> with GetLifeCycle {
   /// Checks whether the controller has already been initialized.
   bool get initialized => _initialized;
 
-  bool _isClosed = false;
-
-  /// Checks whether the controller has already been closed.
-  bool get isClosed => _isClosed;
-
   // Internal callback that starts the cycle of this controller.
   void _onStart() {
     if (_initialized) return;
     onInit();
     _initialized = true;
     SchedulerBinding.instance?.addPostFrameCallback((_) => onReady());
-  }
-
-  // Internal callback that starts the cycle of this controller.
-  void _onDelete() {
-    if (_isClosed) return;
-    _isClosed = true;
-    onClose();
   }
 
   RxStatus _status;
