+++ test/instance/get_instance_test.dart
@@ -12,7 +12,6 @@ class Mock {
 class DisposableController with GetLifeCycle {
   DisposableController() {
     onStart.callback = _onStart;
-    onDelete.callback = _onDelete;
   }
 
   // Internal callback that starts the cycle of this controller.
@@ -21,18 +20,13 @@ class DisposableController with GetLifeCycle {
     onInit();
   }
 
-  // Internal callback that starts the cycle of this controller.
-  void _onDelete() {
-    if (isClosed) return;
-    isClosed = true;
-    onClose();
-  }
-
   bool initialized = false;
 
   bool isClosed = false;
 
+  @override
   void onInit() async {
+    super.onInit();
     initialized = true;
   }
 }
